fix some strong and em commonmark examples failed test
