+++ src/muya/lib/parser/index.js
@@ -1,6 +1,6 @@
 import { beginRules, inlineRules } from './rules'
 import { isLengthEven, union } from '../utils'
-import { punctuation, WHITELIST_ATTRIBUTES } from '../config'
+import { getAttributes, parseSrcAndTitle, validateEmphasize, lowerPriority } from './utils'
 
 // const CAN_NEST_RULES = ['strong', 'em', 'link', 'del', 'a_link', 'reference_link', 'html_tag']
 // disallowed html tags in https://github.github.com/gfm/#raw-html
@@ -10,100 +10,6 @@ delete validateRules.em
 delete validateRules.strong
 delete validateRules['tail_header']
 delete validateRules['backlash']
-
-const validWidthAndHeight = value => {
-  if (!/^\d{1,}$/.test(value)) return ''
-  value = parseInt(value)
-  return value >= 0 ? value : ''
-}
-
-const getAttributes = html => {
-  const parser = new DOMParser()
-  const doc = parser.parseFromString(html, 'text/html')
-  const target = doc.querySelector('body').firstElementChild
-  if (!target) return null
-  const attrs = {}
-  for (const attr of target.getAttributeNames()) {
-    if (!WHITELIST_ATTRIBUTES.includes(attr)) continue
-    if (/width|height/.test(attr)) {
-      attrs[attr] = validWidthAndHeight(target.getAttribute(attr))
-    } else {
-      attrs[attr] = target.getAttribute(attr)
-    }
-  }
-
-  return attrs
-}
-
-const parseSrcAndTitle = (text = '') => {
-  const parts = text.split(/\s+/)
-  const src = parts[0]
-  const rawTitle = text.substring(src.length).trim()
-  const TITLE_REG = /^('|")(.*?)\1$/ // we only support use `'` and `"` to indicate a title now.
-  let title = ''
-  if (rawTitle && TITLE_REG.test(rawTitle)) {
-    title = rawTitle.replace(TITLE_REG, '$2')
-  }
-  return { src, title }
-}
-
-const lowerPriority = (src, offset) => {
-  let i
-  for (i = 0; i < offset; i++) {
-    const text = src.substring(i)
-    for (const rule of Object.keys(validateRules)) {
-      const to = validateRules[rule].exec(text)
-      if (to && to[0].length > offset - i) {
-        return false
-      }
-    }
-  }
-  return true
-}
-
-const validateEmphasize = (src, offset, marker, pending) => {
-  /**
-   * Intraword emphasis is disallowed for _
-   */
-  const lastChar = pending.charAt(pending.length - 1)
-  const followedChar = src[offset]
-  const ALPHA_REG = /[a-zA-Z]{1}/
-  if (/_/.test(marker)) {
-    if (ALPHA_REG.test(lastChar)) return false
-    if (followedChar && ALPHA_REG.test(followedChar)) return false
-  }
-  /**
-   * 1. This is not emphasis, because the second * is preceded by punctuation and followed by an alphanumeric
-   * (hence it is not part of a right-flanking delimiter run:
-   * 2. This is not emphasis, because the opening * is preceded by an alphanumeric and followed by punctuation,
-   * and hence not part of a left-flanking delimiter run:
-   */
-  if (ALPHA_REG.test(lastChar) && punctuation.indexOf(src[marker.length]) > -1) {
-    return false
-  }
-
-  if (followedChar && ALPHA_REG.test(followedChar) && punctuation.indexOf(src[offset - marker.length - 1]) > -1) {
-    return false
-  }
-  /**
-   * When there are two potential emphasis or strong emphasis spans with the same closing delimiter,
-   * the shorter one (the one that opens later) takes precedence. Thus, for example, **foo **bar baz**
-   * is parsed as **foo <strong>bar baz</strong> rather than <strong>foo **bar baz</strong>.
-   */
-  const mLen = marker.length
-  const emphasizeText = src.substring(mLen, offset - mLen)
-  const index = emphasizeText.indexOf(marker)
-  if (index > -1 && /\S/.test(emphasizeText[index + mLen])) {
-    return false
-  }
-  /**
-   * Inline code spans, links, images, and HTML tags group more tightly than emphasis.
-   * So, when there is a choice between an interpretation that contains one of these elements
-   * and one that does not, the former always wins. Thus, for example, *[foo*](bar) is parsed
-   * as *<a href="bar">foo*</a> rather than as <em>[foo</em>](bar).
-   */
-  return lowerPriority(src, offset)
-}
 
 const tokenizerFac = (src, beginRules, inlineRules, pos = 0, top, labels) => {
   const tokens = []
@@ -208,7 +114,8 @@ const tokenizerFac = (src, beginRules, inlineRules, pos = 0, top, labels) => {
     for (const rule of emRules) {
       const to = inlineRules[rule].exec(src)
       if (to && isLengthEven(to[3])) {
-        const isValid = validateEmphasize(src, to[0].length, to[1], pending)
+        console.log(to)
+        const isValid = validateEmphasize(src, to[0].length, to[1], pending, validateRules)
         if (isValid) {
           inChunk = true
           pushPending()
@@ -239,7 +146,7 @@ const tokenizerFac = (src, beginRules, inlineRules, pos = 0, top, labels) => {
     for (const rule of chunks) {
       const to = inlineRules[rule].exec(src)
       if (to && isLengthEven(to[3])) {
-        if (rule === 'emoji' && !lowerPriority(src, to[0].length)) break
+        if (rule === 'emoji' && !lowerPriority(src, to[0].length, validateRules)) break
         inChunk = true
         pushPending()
         const range = {
