+++ test/specs/commonMark/commonmark.0.29.json
@@ -2768,8 +2768,7 @@
     "example": 341,
     "start_line": 6003,
     "end_line": 6007,
-    "section": "Code spans",
-    "shouldFail": true
+    "section": "Code spans"
   },
   {
     "markdown": "[not a `link](/foo`)\n",
@@ -2980,8 +2979,7 @@
     "example": 367,
     "start_line": 6455,
     "end_line": 6459,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "*(*foo*)*\n",
@@ -3014,8 +3012,7 @@
     "example": 371,
     "start_line": 6497,
     "end_line": 6501,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "_(_foo_)_\n",
@@ -3080,8 +3077,7 @@
     "example": 379,
     "start_line": 6570,
     "end_line": 6574,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "foo**bar**\n",
@@ -3180,8 +3176,7 @@
     "example": 391,
     "start_line": 6685,
     "end_line": 6689,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "*(**foo**)*\n",
@@ -3229,8 +3224,7 @@
     "example": 397,
     "start_line": 6742,
     "end_line": 6746,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "_(__foo__)_\n",
@@ -3246,8 +3240,7 @@
     "example": 399,
     "start_line": 6761,
     "end_line": 6765,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "__пристаням__стремятся\n",
@@ -3886,8 +3879,7 @@
     "example": 475,
     "start_line": 7385,
     "end_line": 7389,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "__<a href=\"__\">\n",
@@ -3895,8 +3887,7 @@
     "example": 476,
     "start_line": 7392,
     "end_line": 7396,
-    "section": "Emphasis and strong emphasis",
-    "shouldFail": true
+    "section": "Emphasis and strong emphasis"
   },
   {
     "markdown": "*a `*`*\n",
